pyproject.toml
[tool.poetry]
name = "practiceprogramingpython"
version = "1.0.0-Alpha"
description = ""
authors = ["DimaL007"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "==0.135.2"
uvicorn = "^0.41.0"
email-validator = "^2.3.0"
myapplication = "^0.1.0"
the-new-hotness = "^1.3.0"
django-email-users = "^2.1.0"
pydantic2 = "^1.1.15"
poetry = "^2.3.2"
sqlalchemy = "^2.0.48"
asyncpg = "^0.31.0"
alembic = "^1.18.4"
pydantic-settings = "^2.13.1"
python-dotenv = "^1.2.2"
async = "^0.6.2"
automium-web = "^0.1.1"
core-models = "^0.5.7"
python-jose = "^3.5.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
